requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
PyYAML==6.0.1
pyarrow==14.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
